README: move and make it generic for TI and NXP sub-layers
Both layers will have very similar content and usage so it's best to
have a single README file that can be used for both.

Move it into the root of the meta-optologic layer and rename it to
README-tdx.md to indicate that it contains information for both TI and
NXP sub-layers for Toradex SoMs.

Change-Id: Iabb4f143720a431f9937ac84bbf89c50ab25a272